+++ doc/index.rst
@@ -15,6 +15,8 @@ Contents:
 
    cholmod.rst
 
+   .. sksparse._cholmod_internal <_cholmod_internal.rst>
+
    changes.rst
 
    Python Module Index <modindex>
