+++ setup.py
@@ -83,7 +83,7 @@ setup(
         "": ["test_data/*.mtx.gz"],
     },
     name=DISTNAME,
-    version="0.4.16",  # remember to update __init__.py
+    version="0.5.0",  # remember to update __init__.py
     maintainer=MAINTAINER,
     maintainer_email=MAINTAINER_EMAIL,
     description=DESCRIPTION,
