+++ sksparse/__init__.py
@@ -1,1 +1,1 @@
-__version__ = "0.4.16"
+__version__ = "0.5.0"
